feat: Populate MenuItemCategoryActivity's RecyclerView using SubCategoryAdapter
Clicking "Hot Coffees" from MenuFragment's RecyclerView of Category starts MenuItemCategoryActivity, which has a collapsing toolbar and a RecyclerView of SubCategory (each list item of SubCategory has a TextView [for its title] and a nest-RecyclerView [displaying a horizontally-scrollable list of MenuItem representing the SubCategory]).

The MenuItem within the nested-RecyclerView are responding to clicks and long-clicks by displaying a toast-message.

NEXT: create a DetailActivity that loads a clicked MenuItem so users can view and customize the concrete classes of Drink.