refactor: Rework MainActivity's view hierarchy to not have NestedScrollView
The main content of MainActivity's view hierarchy is now simply a FragmentContainerView. The layout of the Fragments hosted in this FCV will each have (either) the NestedScrollView (or a RecyclerView [which already has nested scrolling]).

A key part of fixing MenuItemCategoryActivity's collapsing toolbar un-smoothness was calling the following on the nested-RV: "setNestedScrollingEnabled(false)".