fix: "Up" button no longer recreating parent activity
Set the launchMode of the parent activity to be "singleTop" and set a flag on the "up" intent to "FLAG_ACTIVITY_CLEAR_TOP".

"Similarly, if the user [navigates up] to an activity on the current stack, the behavior is determined by the parent activity's launch mode. If the parent activity has launch mode [singleTop] (or the [up] intent contains [FLAG_ACTIVITY_CLEAR_TOP]), the parent is brought to the top of the stack, and its state is preserved."

Also, changed the layout for the RecyclerView displaying MenuItem so that the name of the MenuItem (displayed in a TextView) will be layout on multiple lines (in this case, defined as 2 lines).