tinker: Change "Up" button icon in MenuItemActivity
Used Toolbar.setNavigationIcon() to change the "Up" button to an "x" icon, then override the behavior of onOptionsItemSelected() when the MenuItem's itemId is "android.R.id.home" to call finish() on this detail activity.

Also, added a toolbar that collapses/expands on vertical scroll events (similar to the one from MainActivity) that displays the title of the selected MenuItem.