refactor: Modify vertical spacing for list items
Cosmetic changes to allow each list item to have more vertical spacing, making it easier for user to make a selection.

NEXT: Implement the "customize" button (start a new activity with more customization options for the drink that we are currently building [the drink may already be partially customized at the point of starting this new activity]).