tinker: Add RecyclerView and Button to layout
Button's click handler does the same as the "home" button's custom response (it calls finish() on the Activity). Currently, the RecyclerView does not have an adapter or layout manager attached to it.

Notes left in class CaffeLatte about different kinds of Incrementable (those that track 1 quantity and those that must track multiple quantities). Also, thinking about how to distinguish between DrinkComponent Type(s) that are ALLOWABLE and those that are NOT ALLOWABLE for specific concrete Drink classes.