tinker: Utilize null-checking for PullOptions and PrepOptions
Drink.drinkComponentsDefaultAsString should not use a loop that depends on what's been put into Drink.drinkComponents (sometimes it'll have DrinkComponent with types set to [null]... which isn't the value we need for the collection of default values [we need something for the "DEFAULT" indicator]). Instead, create a list of default values from the default-constants defined in CaffeLatte.

Not seeing PullOptions/PrepOptions listed in MenuItemActivity (which is the correct behavior). PullOptions/PrepOptions are showing up in CustomizeActivity (which is the correct behavior), but with visual and behaviorial bugs (they're being displayed as customized [green border with title] when they should appear as a third type [not INCREMENTABLE nor SINGLE_SELECTION], one without a border and with a greyed-out background) (the behavior so far is to select a type, which is just part of what we need... we need it to add/remove from Drink.drinkComponents).