tinker: Set up item click handler for DrinkComponentIncrementableViewHolder
Differentiate between Incrementables with one vs many Type.

Those with one Type value should NOT do anything when the list item is clicked. Instead, it should allow its ivMinus's and ivAdd's click handlers to modify the Incrementable's quantity.

Those with many Type values should show a ModalBottomSheet with a list of the many Type values. When selected, a new DrinkComponent (with a non-null type and quantity set to 1) should be added to the list. These do NOT get removed from the list even when their quantity is 0 (they'll just display their initial looks [gray borderless, titleless background with some of its views invisible]). Must track which of the multiple types have been added to the list and which haven't (the ModalBottomSheet should NOT list types that have already been added to the list).

Getting ready to implement FlavorOptions and SweetenerOptions, both of which are Incrementable with multiple Type values.