refactor: Override DrinkComponentBaseAdapter view holders' click handlers
Incrementables with multiple options for Type should display a bottom sheet to select a Type, then add a new DrinkComponent to the list after selection.

BUGGY: The new DrinkComponent is being added to the list, but it's NOT displaying as intended. Clicking the new DrinkComponent's ivAdd does NOT increment quantity, instead it reverts the looks of the DrinkComponent to its initial version (gray borderless background).