tinker: Add Drizzle to CaffeLatte
Drizzle is a Granular with multiple values for its Type. Looks/behaviors are good for MenuItemActivity.

BUGGY: CustomizeActivity should remove the Granular if the default value is selected (but not when it's (1) a Granular with a single value for its Type or (2) the last of a Granular with multiple values for its Type [need it to become an "invoker"]).