fix: Remove duplicate Granular "invoker" in CustomizeAdapter constructor
The overall BUG from before revolves around not updating the underlying Drink model (instead, WhatsIncludedAdapter uses a list that is compiled using the underlying Drink model... this list gets updated but does not update the underlying Drink model that was used to generate it). I could redo the source of WhatsIncludedAdapter so that changes made will get reflected for the underlying Drink model, BUT it's actually very complicated to translate it back from the compiled list into the separate lists used to generate it.

HEAR ME OUT: for WhatsIncludedAdapter (MenuItemActivity), the only updates to Granular are selecting a different amount value or its default value (causing a removal)... the only updates needed to syncronize the data are removal-updates (which led to this version of filtering out duplicates).