tinker: Add CupOptions to CaffeLatte
CupOptions has one concrete subclass: CupSize. CupSize is part of the standard recipe for CaffeLatte (with its default value being: NO).

It is a special case of a drink component being a part of the standard recipe and starting off with a defined type, but does NOT appear in MenuItemActivity unless a non-default type has been selected from CustomizeActivity (where it does appear). Also, selecting the default while in MenuItemActivity will remove it (even though it's part of the standard recipe).