tinker: Add LineTheCup to CaffeLatte
LineTheCup is an AddInsOptions that (like CupSize) is part of the standard recipe, but is not displayed in MenuItemActivity if its default value is selected.

NEXT: Add the last AddInOptions for CaffeLatte (Powders). Powders is unique (previously, all the types of a particular DrinkComponent were all the same [all Incrementable, all Granular, or all simple selection]) (now, some of Powders's types are Incrementable while other types are Granular).

NEXT: Implement changing drink-size.