fix: Scroll content of CustomizeActivity instead of just RV
Nest CustomizeActivity's RV (and the rest of CustomizeActivity's content) into a NestedScrollView and call setNestedScrollingEnabled(false) on the RV. Now, the drink banner at the top and the "Done customizing" button at the bottom will scroll with the RV inbetween them. Previously, the user was only able to scroll the RV while the drink banner and "Done customizing" button remained stationary on the screen. Padding on the bottom of certain View(s) was really important in minimizing overlap with the FAB and "Cart" button.

This version is visually more aligned with the siren's app.