tinker: Implement "Add to order" and "Cart" buttons
OrderFragment is now a singleton that holds the order (List<MenuItem>). "Add to order" button from MenuItemActivity and CustomizeActivity will add the current drink to OrderFragment's order. "Cart" button from MenuItemActivity and CustomizeActivity will start ReviewOrderActivity.

ReviewOrderActivity has a collapsing toolbar whose title is partially derived from the size of the order passed into it. The RV for this activity is almost finished, but it needs a nested RV to display the drink's customizations and implementation to the add and minus buttons.