tinker: Add Fruits to CaffeLatte (will remove later)
Also, fix view holder's drop-down icon for "invokers" that are [MixedType] (which uses ViewHolderSimpleSelection).

Setting up to account for the number of [MixedType] drink components that are already in the drink compared to the number of [MixedType] types to select from. This will be performed in CustomizeInnerAdapter.handleSelectionOfEverythingElse(), within the if-else clause for [MixedType].

NOTE: Any drink component that implement [MixedType] shall have at least 2 [MixedType] types to select from (at least 1 Incrementable and at least 1 Granular). This is important when the "invoker" needs to be substituted for a concrete drink component.