tinker: Replace Fruits "invoker" w last option
This approach adds the user selection BEFORE it checks whether the "invoker" should be replaced (e.g. one option left AFTER user selection).

The previous approach (e.g. Granular) checks for second-to-last condition, finds the last 2 types NOT inside the drink, and determines which of the 2 was the one selected by the user and which is the last option. This previous approach sets the user selection AFTER checking condition for "invoker" replacement.

NEXT: filter the list given to bottom modal sheet (do not show Fruits drink component options that are already inside the drink).