refactor: Compete with RV for focus
Instead of blocking focus on descendants (which could cause trouble when there's an EditText inside the RV item's layout), have RV's parent compete for focus ("As I see in NestedScrollView source it tries to focus the first possible child in onRequestFocusInDescendants and if only RecyclerView is focusable it wins.").

Also, set RV's nested scrolling to false since it's inside a NestedScrollView.

REF: https://stackoverflow.com/questions/36314836/recycler-view-inside-nestedscrollview-causes-scroll-to-start-in-the-middle