fix: Use instance variable [quantityMin] to define default value
Previously, used class level constant.

Each drink should be able to define its own minimum value for a drink component (e.g. CaffeLatte's Shot should be defined with a minimum quantity of 1 [otherwise it's not a latte, it's steamed milk], while drinks with optional Shot [brewed coffee with an add Shot] should be defined with a minimum quantity of 0).