fix: Factor in standard recipe for WhatsIncludedAdapter.generateWhatsIncluded(Drink)
Re-work logic for generating the list of drink components displayed in MenuItemActivity's RV.

Previously, the approach was to remove all "invoker" and drink components with quantity/amount representing default, and include whatever's left.

Now, include it if it's in the standard recipe (except for CupSize and LineTheCup with default), and include allowables that do not represent default.